move contents of pkg/util/test into pkg/test
