+++ pkg/backup/item_hook_handler_test.go
@@ -32,8 +32,8 @@ import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
 
 	v1 "github.com/heptio/velero/pkg/apis/velero/v1"
+	velerotest "github.com/heptio/velero/pkg/test"
 	"github.com/heptio/velero/pkg/util/collections"
-	velerotest "github.com/heptio/velero/pkg/util/test"
 )
 
 type mockItemHookHandler struct {
