+++ pkg/backup/pod_action_test.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/heptio/velero/pkg/kuberesource"
 	"github.com/heptio/velero/pkg/plugin/velero"
-	velerotest "github.com/heptio/velero/pkg/util/test"
+	velerotest "github.com/heptio/velero/pkg/test"
 )
 
 func TestPodActionAppliesTo(t *testing.T) {
