+++ pkg/plugin/clientmgmt/client_builder_test.go
@@ -26,7 +26,7 @@ import (
 	"github.com/stretchr/testify/assert"
 
 	"github.com/heptio/velero/pkg/plugin/framework"
-	"github.com/heptio/velero/pkg/util/test"
+	"github.com/heptio/velero/pkg/test"
 )
 
 func TestNewClientBuilder(t *testing.T) {
