+++ pkg/plugin/framework/backup_item_action_test.go
@@ -32,7 +32,7 @@ import (
 	"github.com/heptio/velero/pkg/backup/mocks"
 	proto "github.com/heptio/velero/pkg/plugin/generated"
 	"github.com/heptio/velero/pkg/plugin/velero"
-	velerotest "github.com/heptio/velero/pkg/util/test"
+	velerotest "github.com/heptio/velero/pkg/test"
 )
 
 func TestBackupItemActionGRPCServerExecute(t *testing.T) {
