+++ pkg/restore/add_pvc_from_pod_action_test.go
@@ -28,7 +28,7 @@ import (
 
 	"github.com/heptio/velero/pkg/kuberesource"
 	"github.com/heptio/velero/pkg/plugin/velero"
-	velerotest "github.com/heptio/velero/pkg/util/test"
+	velerotest "github.com/heptio/velero/pkg/test"
 )
 
 func TestAddPVCFromPodActionExecute(t *testing.T) {
